feat(/client, /server): Added first functional blog system.
I used Notion API to create my posts. I get data from my Notion database with request in the server part of the project. In the client part (front), I used axios to make request on the server part. I created a style for post and summaryCard in article. I added a loading 'spinner' for more UX design (because get from API is long, so we can improve that, in another version of the project.