Feature/blog : Created all blog component + pages + connection with Notion API (#1)
* feat(src/):Change structure of blog + added components

Now, we have a component categoryDetails who get path of url and display information.

* feat(/client, /server): Added first functional blog system.

I used Notion API to create my posts. I get data from my Notion database with request in the server part of the project. In the client part (front), I used axios to make request on the server part. I created a style for post and summaryCard in article. I added a loading 'spinner' for more UX design (because get from API is long, so we can improve that, in another version of the project.

* feat(/client, /server/notionIntegration.ts): Added categories in navmenu + added filters

I removed useless commented codes. I added categories of articles (in tags). This categories are added in the NavMenu component. When we click on a categories, there are a filter to just display corresponding articles

* chore(client/): Clean code (removed console.log + class to className)

I removed all useless console.log in blog section. I changed class to className. I added key props on element in a 'map' loop.