Feature/about : I added server files (#3)
* feat(includes/, maint-content/, css/): Added QuickSlideTab and QuickSlideButton components

I added some new components to create an quick access to element. The button can get a link to go on a particular page

* feat(client/src): Added AboutMe page and CV page

I created 2 components : AboutMe who is a component displayed when we go on /aboutme. CurriculumVitae who is a component displayed when we go on /cv

* feat(client/, server/): Added formation and experiences pages.

I added style for formation and experiences pages + I added a data.json file to get data from this file (data for formation, experiences and competences). It's a easy way to organize my data.

* feat(client/): Added some style for all elements of about pages

* feat(client/, server/): I created the company page component to get information about companies

Now, when we click on the name of company (in Experience), we go on a page with information about the company.

* feat(server/, client/): Changed Competence components + Added new company page (freelance)

* feat(server/): I pushed server files (because I forgot to push it)