Hotfix/blog : Fixed problem with link of post url + center images in post (#12)
* hotfix(Post.js): I added a second way to get post (with use of the slug in the url)

I added this method because when we clicked on a link to a post (but not in /blog page) we get an error (because the request doesn't get the id in url). So I just added a way to get the url in the url (who is the slug). To to that, I used regex.

* hotfix(Blog.css): I added property to center image